Create login api route - installed jsonwebtoken - added login functionalities - remove password from response - fixed authRoutes path Changes to be committed: 	modified:   package-lock.json 	modified:   package.json 	modified:   server/controllers/auth.controller.js 	renamed:    server/middlewares/error.js -> server/middleware/error.js 	modified:   server/routes/auth.route.js 	modified:   server/server.js
