+++ package.json
@@ -13,6 +13,7 @@
     "bcryptjs": "^2.4.3",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.3.0",
     "nodemon": "^3.1.0"
   }
