+++ server/server.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const mongoose = require("mongoose");
 const dotenv = require('dotenv')
-
-const authRoutes = require('./controllers/auth.controller');
+const authRoutes = require('./routes/auth.route');
 dotenv.config();
 
 mongoose
